users can log in after token expires

+++ cli/requests/test.go
@@ -8,7 +8,7 @@ import (
 	"google.golang.org/grpc/metadata"
 )
 
-func Test(conn *grpc.ClientConn, token string) (*int32, error) {
+func Test(conn *grpc.ClientConn, token string) (*int64, error) {
 
 	newMetadata := metadata.Pairs("Authorization", token)
 	newContext := metadata.NewOutgoingContext(context.Background(), newMetadata)
